rework and additions, moved filed
