+++ client/src/components/pages/admin/admin.component.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import './admin.component.css';
+import { NavPanelComponent } from './nav-panel/nav-panel.component';
 import { AllTicketsComponent } from './all/all-tickets.component';
 import { RecentTicketsComponent } from './recent/recent-ticket.component';
 import { AcceptedTicketsComponent } from './accepted/accepted-tickets.component';
@@ -8,17 +9,19 @@ import { AcceptedTicketsComponent } from './accepted/accepted-tickets.component'
 export const AdminComponent: React.FC = () => {
     
     return (
-        <div>
+        <div className='primary'>
 
-            <nav> NavPanel Here</nav>
+            <div className='nav'>
+                <NavPanelComponent />
+            </div>
             
-            <main>
+            <div className='main'>
                 <RecentTicketsComponent />
 
                 <AcceptedTicketsComponent />
 
                 <AllTicketsComponent />
-            </main>
+            </div>
         </div>
     );
 };
