+++ client/src/components/pages/admin/nav-panel/nav-panel.component.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import './nav-panel.component.css';
+
+//NavPanel Here
+export const NavPanelComponent: React.FC = () => {
+
+    return (
+
+        <div className='centerize'>
+            <a href='./'>
+                <div className='button'>
+                    New Ticket
+                </div>
+            </a>
+        </div>
+                
+        
+    );
+};
